requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
supabase
pillow
python-multipart
exponent-server-sdk